add mais metodos de segurança

+++ frontend/src/pages/HomePage/index.js
@@ -1,13 +1,41 @@
 import React from "react";
+import axios from "axios"; // Não esqueça de importar o Axios
 import { removeToken } from "../../services/auth";
+import { ToastContainer, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 import "./styles.css";
-// import io from "socket.io-client";
 
 function HomePage() {
   const handleLogout = () => {
-    removeToken(); // Remove o token
-    localStorage.clear(); //limpa o localstor
-    window.location.reload(); // Recarrega a página
+    // Envia uma solicitação POST para a rota de logout no servidor
+    const token = localStorage.getItem("token"); // Supondo que o token está armazenado no Local Storage
+
+    axios
+      .post(
+        "http://localhost:3001/logout",
+        {},
+        {
+          headers: {
+            Authorization: `Bearer ${token}`, // Adiciona o token ao header da solicitação
+          },
+        }
+      )
+      .then((response) => {
+        if (response.status === 200) {
+          toast.success("Logout realizado com sucesso!", {
+            autoClose: 100,
+            onClose: () => {
+              removeToken(); // Remove o token
+              localStorage.clear(); //limpa o localstor
+              window.location.reload(); // Recarrega a página
+            },
+          });
+        }
+      })
+      .catch((error) => {
+        console.error(error);
+        alert("Erro ao fazer logout");
+      });
   };
 
   const redirectCadastro = () => {
@@ -34,6 +62,7 @@ function HomePage() {
             value="Logout"
             onClick={handleLogout}
           />
+          <ToastContainer />
         </div>
       </div>
       <div className="content">
